1.refactored the project structure 2.updated dependency 3.fixed the SwitchToLogin button usecase bug
