--- src/app/ServerApp.java
@@ -1,100 +1,0 @@
-package app;
-
-import data_access.FileGroupDataAccessObject;
-import data_access.FileSaveDataAccessObject;
-import data_access.FileSummayDataAccessObject;
-import data_access.FileUserDataAccessObject;
-import entity.CommonGroupFactory;
-import entity.CommonUserFactory;
-import interface_adapter.clear_users.ClearViewModel;
-import interface_adapter.login.LoginViewModel;
-import interface_adapter.logged_in.LoggedInViewModel;
-import interface_adapter.notify.NotifyViewModel;
-import interface_adapter.signup.SignupViewModel;
-import interface_adapter.ViewManagerModel;
-import view.LoggedInView;
-import view.LoginView;
-import view.SignupView;
-import view.ViewManager;
-
-import javax.swing.*;
-import java.awt.*;
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-
-public class ServerApp {
-    public static void main(String[] args) throws IOException {
-
-
-
-        // Build the main program window, the main panel containing the
-        // various cards, and the layout, and stitch them together.
-
-        // The main application window.
-        JFrame application = new JFrame("Login Example");
-        application.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-
-        CardLayout cardLayout = new CardLayout();
-
-        // The various View objects. Only one view is visible at a time.
-        JPanel views = new JPanel(cardLayout);
-        application.add(views);
-
-        // This keeps track of and manages which view is currently showing.
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        new ViewManager(views, cardLayout, viewManagerModel);
-
-        // The data for the views, such as username and password, are in the ViewModels.
-        // This information will be changed by a presenter object that is reporting the
-        // results from the use case. The ViewModels are observable, and will
-        // be observed by the Views.
-        LoginViewModel loginViewModel = new LoginViewModel();
-        LoggedInViewModel loggedInViewModel = new LoggedInViewModel();
-        SignupViewModel signupViewModel = new SignupViewModel();
-        ClearViewModel clearViewModel = new ClearViewModel();
-        NotifyViewModel notifyViewModel = new NotifyViewModel();
-
-        FileUserDataAccessObject userDataAccessObject;
-        try {
-            userDataAccessObject = new FileUserDataAccessObject("./users.csv", new CommonUserFactory());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        FileSummayDataAccessObject summaryUserDataAccessObject;
-        summaryUserDataAccessObject = new FileSummayDataAccessObject();
-
-        FileSaveDataAccessObject saveUserDataAccessObject;
-        saveUserDataAccessObject = new FileSaveDataAccessObject();
-
-        FileGroupDataAccessObject groupDataAccessObject;
-        try {
-            groupDataAccessObject = new FileGroupDataAccessObject("./groups.csv", new CommonGroupFactory());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        SignupView signupView = SignupUseCaseFactory.create(viewManagerModel, loginViewModel, signupViewModel,
-                userDataAccessObject, userDataAccessObject, clearViewModel);
-        views.add(signupView, signupView.viewName);
-
-        LoginView loginView = LoginUseCaseFactory.create(viewManagerModel, loginViewModel, loggedInViewModel,
-                signupViewModel, userDataAccessObject, groupDataAccessObject);
-        views.add(loginView, loginView.viewName);
-
-        LoggedInView loggedInView = LoggedInUseCaseFactory.create(viewManagerModel,
-                loginViewModel,
-                loggedInViewModel,
-                userDataAccessObject,
-                summaryUserDataAccessObject, saveUserDataAccessObject);
-
-        views.add(loggedInView, loggedInView.viewName);
-
-        viewManagerModel.setActiveView(signupView.viewName);
-        viewManagerModel.firePropertyChanged();
-
-        application.pack();
-        application.setVisible(true);
-    }
-}
