+++ src/app/factories/LoggedInUseCaseFactory.java
@@ -1,4 +1,4 @@
-package app;
+package app.factories;
 
 import entity.CommonUserFactory;
 import entity.UserFactory;
