+++ src/interface_adapter/signup/SignupViewModel.java
@@ -16,7 +16,7 @@ public class SignupViewModel extends ViewModel {
 
     public static final String SIGNUP_BUTTON_LABEL = "Sign up";
     public static final String CANCEL_BUTTON_LABEL = "Cancel";
-    public static final String SWITCH_TO_LOGIN_BUTTON_LABEL = "Login";
+    public static final String SWITCH_TO_LOGIN_BUTTON_LABEL = "Switch to Login";
 
     private SignupState state = new SignupState();
